# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hull Local API
  version: "1.0.0"
  description: |
    The contract between `hulld` and its clients (CLI, and the native GUIs from
    Phase 3 Wave 2 onward). This document is the FROZEN contract: every native
    client builds against it. Keep it in sync with `internal/api` , the
    coverage test `internal/api/openapi_test.go` fails if a registered route is
    missing here.

    Transport (ADR 0006): HTTP on `127.0.0.1:<port>` where `<port>` and a bearer
    token are read from `~/.hull/daemon.json` (written 0600 by a running
    daemon). Auth accepts `Authorization: Bearer <token>` or `?token=<token>`
    (the latter for EventSource/SSE, which cannot set headers). The server
    reflects any Origin for CORS , safe because the token, not the origin, is
    the access control and it only listens on loopback.

    Long operations (create, import, add-service, link, rebuild, reset) are
    JOBS: a POST returns a job reference; observe it by polling `GET
    /v1/jobs/{id}` or streaming `GET /v1/jobs/{id}/stream`.
servers:
  - url: http://127.0.0.1:{port}
    description: Local daemon (port from ~/.hull/daemon.json)
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /v1/projects:
    get:
      tags: [projects]
      summary: List registered projects with live running state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectInfo" }
    post:
      tags: [projects]
      summary: Create a project (job)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectRequest" }
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
  /v1/projects/{name}/{action}:
    post:
      tags: [projects]
      summary: Lifecycle action on a project
      description: |
        `start`, `stop`, `restart`, and `repair` are synchronous (204).
        `rebuild` and `reset` run as jobs (202). `restart` force-recreates
        containers; `repair` brings the project down (keeping volumes) and back
        up. `reset` removes named volumes. `?no_cache=1` on `rebuild`.
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, restart, repair, rebuild, reset]
        - { name: no_cache, in: query, required: false, schema: { type: boolean } }
      responses:
        "204": { description: Done (start/stop/restart/repair) }
        "202":
          description: Job started (rebuild/reset)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
  /v1/projects/{name}/volumes:
    get:
      tags: [projects]
      summary: List a project's named volumes (Reset blast radius)
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { type: string } }
...
  /v1/imports:
    post:
      tags: [projects]
      summary: Import an unmanaged folder or legacy project under a root (job)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportRequest" }
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
  /v1/clusters:
    get:
      tags: [clusters]
      summary: List adopted/managed clusters (reconciled with the started ledger)
      responses:
        "200":
          description: Clusters
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClusterInfo" }
    post:
      tags: [clusters]
      summary: Adopt an existing compose project as a Hull cluster
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdoptClusterRequest" }
      responses:
        "200":
          description: Adopted
          content:
            application/json:
              schema:
                type: object
                properties: { name: { type: string } }
  /v1/clusters/{name}:
    put:
      tags: [clusters]
      summary: Set a cluster's base_domain and/or ingress mode
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetClusterConfigRequest" }
      responses:
        "204": { description: Updated }
  /v1/clusters/{name}/routes/{key}:
    put:
      tags: [clusters]
      summary: Assign a URL (subdomain) to one of the cluster's services
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: key, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetRouteRequest" }
      responses:
        "204": { description: Route assigned }
    delete:
      tags: [clusters]
      summary: Remove a cluster route by its subdomain key
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: key, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Route removed }
  /v1/clusters/create:
    post:
      tags: [clusters]
      summary: Create a new multi-container project (job)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClusterRequest" }
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
...
  /v1/jobs/{id}:
    get:
      tags: [jobs]
      summary: One job snapshot
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "404": { description: Unknown job, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorBody" } } } }
  /v1/jobs/{id}/stream:
    get:
      tags: [jobs]
      summary: Stream a job's log lines (SSE)
      description: "text/event-stream. Each `data:` is a log line; ends with `event: done` carrying the final snapshot."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: SSE stream, content: { text/event-stream: { schema: { type: string } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    tokenQuery:
      type: apiKey
      in: query
      name: token
  schemas:
...
    ProjectInfo:
      type: object
      required: [name, dir, kind, running, served]
      properties:
        name: { type: string }
        dir: { type: string }
        kind: { type: string, description: "template name, app, legacy, folder, or invalid" }
        url: { type: string }
        running: { type: boolean }
        legacy: { type: boolean }
        error: { type: string }
        php: { type: string }
        served: { type: boolean }
        group: { type: string }
        services: { type: array, items: { $ref: "#/components/schemas/ProjectServiceInfo" } }
        routes: { type: array, items: { $ref: "#/components/schemas/ClusterRouteInfo" } }
    ProjectServiceInfo:
      type: object
      required: [key, engine, mode]
      properties:
        key: { type: string }
        engine: { type: string }
        version: { type: string }
        mode: { type: string, enum: [dedicated, shared] }
        instance: { type: string }
    ClusterRouteInfo:
      type: object
      required: [key, subdomain, service, port, served]
      properties:
        key: { type: string }
        subdomain: { type: string }
        service: { type: string }
        port: { type: integer }
        served: { type: boolean }
        aliases: { type: array, items: { type: string } }
        hosts: { type: array, items: { type: string } }
    SetRouteRequest:
      type: object
      required: [service, port]
      properties:
        service: { type: string }
        port: { type: integer }
        aliases: { type: array, items: { type: string } }
        serve: { type: boolean }
    SetClusterConfigRequest:
      type: object
      properties:
        base_domain: { type: string }
        ingress: { type: string, enum: ["", none, delegate, hull] }
    ClusterInfo:
      type: object
      required: [name, dir, running]
      properties:
        name: { type: string }
        dir: { type: string }
        compose_root: { type: string }
        running: { type: boolean }
        base_domain: { type: string }
        ingress: { type: string, enum: ["", delegate, hull] }
        routes: { type: array, items: { $ref: "#/components/schemas/ClusterRouteInfo" } }
...
    CreateProjectRequest:
      type: object
      required: [name, template]
      properties:
        name: { type: string }
        template: { type: string }
        db: { type: string }
        db_version: { type: string }
        redis: { type: boolean }
        php: { type: string }
        version: { type: string }
        serve: { type: boolean, nullable: true }
        skip_start: { type: boolean }
    AdoptClusterRequest:
      type: object
      required: [dir]
      properties:
        dir: { type: string }
        name: { type: string }
        compose_root: { type: string }
        compose_files: { type: array, items: { type: string } }
        profiles: { type: array, items: { type: string } }
    CreateClusterRequest:
      type: object
      required: [name, containers]
      properties:
        name: { type: string }
        root: { type: string }
        compose_root: { type: string }
        managed: { type: boolean }
        containers: { type: array, items: { $ref: "#/components/schemas/ClusterContainerSpec" } }
    ClusterContainerSpec:
      type: object
      required: [name]
      properties:
        name: { type: string }
        template: { type: string }
        image: { type: string }
        version: { type: string }
        port: { type: integer }
        serve: { type: boolean }
        services: { type: array, items: { $ref: "#/components/schemas/ClusterCardServiceSpec" } }
    ClusterCardServiceSpec:
      type: object
      required: [engine]
      properties:
        engine: { type: string }
        version: { type: string }
    ImportRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    Job:
      type: object
      required: [id, kind, status, lines, created]
      properties:
        id: { type: string }
        kind: { type: string }
        status: { type: string, enum: [running, done, failed] }
        error: { type: string }
        lines: { type: array, items: { type: string } }
        created: { type: string, format: date-time }
    JobRef:
      type: object
      required: [job]
      properties:
        job: { $ref: "#/components/schemas/Job" }
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error: { type: string }